Making a new order now works as expected

+++ mvcJJMS/Models/CartaoCredito.cs
@@ -1,7 +1,7 @@
 namespace mvcJJMS.Models{
 	public class CartaoCredito {
 		//[DatabaseGenerated(DatabaseGeneratedOption.None)] //Prevent the database from automatically generating the primary key for this entity
-		public int CartaoCreditoID{get;set;}
+		public long CartaoCreditoID{get;set;}
 		public int mes{get;set;}
 		public int ano{get;set;}
 		public int cvv{get;set;}
